console: keep the read layer go-mysql-free (#528 guard) — move the flashback protocol server to cmd/bintrail-console
CI's `unit` job failed on TestReadLayerDoesNotLinkGoMySQL (#528): the
embedded flashback server put go-mysql + internal/shim imports into
internal/console, which is a read-layer package barred from linking the
protocol/capture libraries.

Split it:
- internal/console/flashback.go now holds ONLY a go-mysql-free seam:
  FlashbackTarget + (*Server).ResolveFlashback (username -> per-source
  index + baseline), flashbackTarget, flashbackDefaultSchema,
  splitBaselineSource. No go-mysql, no shim import.
- cmd/bintrail-console/flashback.go (package main, already links go-mysql
  via the stream) holds the protocol server: serveFlashback, the accept
  loop, routingHandler, flashbackCreds, bindFlashbackHandler — consuming
  the console seam via ResolveFlashback/Token.
- Tests split accordingly; the wire-routing integration test moved to
  cmd/bintrail-console. watch.go calls the local serveFlashback.

Verified: `go list -deps ./internal/console` now has 0 go-mysql deps; the
#528 guard passes; full unit suite green; the routing integration test
still passes live (routes by id + name, decoy-schema precedence,
wrong-password + unknown-server rejection).
